Move class decorators into one file
